Add trie/util/nibbles module to trie exports
ethereumjs/trie's WalkController exposes the user to the nibbles data structure, but a conversion between buffers and nibbles is not exported in the current version of the library. It leads to developers having to vendor in that code themselves instead of being able to import it from the library that they're anyways installing. This fix adds the functions to the package's exported module.